+++ htrest/apis/time_prog.py
@@ -26,7 +26,7 @@ from htheatpump.httimeprog import TimeProgram as HtTimeProg
 from htheatpump.httimeprog import TimeProgEntry as HtTimeProgEntry
 from .utils import HtContext
 from htrest.app import ht_heatpump
-from htrest.settings import READ_ONLY as HTREST_READ_ONLY
+from htrest import settings
 
 
 _logger = logging.getLogger(__name__)
@@ -97,7 +97,7 @@ class TimeProg(Resource):
             time_prog = ht_heatpump.get_time_prog(id, with_entries=False).as_json(with_entries=False)
             time_prog.update({"entries": api.payload["entries"]})
             time_prog = HtTimeProg.from_json(time_prog)
-            if not HTREST_READ_ONLY:
+            if not settings.READ_ONLY:
                 time_prog = ht_heatpump.set_time_prog(time_prog)
         return time_prog.as_json(with_entries=True)
 
@@ -122,6 +122,6 @@ class TimeProgEntry(Resource):
         _logger.debug("*** {!s} -- id={}, day={}, num={}, payload={!s}".format(request.url, id, day, num, api.payload))
         entry = HtTimeProgEntry.from_json(api.payload)
         with HtContext(ht_heatpump):
-            if not HTREST_READ_ONLY:
+            if not settings.READ_ONLY:
                 entry = ht_heatpump.set_time_prog_entry(id, day, num, entry)
         return entry.as_json()