+++ htrest/settings.py
@@ -27,4 +27,5 @@ RESTX_MASK_SWAGGER = False
 RESTX_ERROR_404_HELP = False
 RESTX_BUNDLE_ERRORS = True
 
-READ_ONLY = False
+BOOL_AS_INT = False  # boolean values are treated as integers (with false equivalent to 0 and true equivalent to 1)
+READ_ONLY = False    # no write accesses to the heat pump; if you want to be sure, that nothing will be manipulated